+++ src/collective/solr/search.py
@@ -9,7 +9,7 @@ from zope.component import queryUtility
 from zope.interface import implementer
 
 from collective.solr.exceptions import SolrInactiveException
-from collective.solr.interfaces import ISearch, ISolrConnectionManager
+from collective.solr.interfaces import ISearch, ISolrConnectionManager, MAX_RESULTS_SUPPORTED_BY_SOLR
 from collective.solr.mangler import (
     cleanupQueryParameters,
     mangleQuery,
@@ -74,13 +74,13 @@ class Search(object):
         parameters["lowercaseOperators"] = lowercase_operator
         parameters["q.op"] = default_operator
         if "rows" not in parameters:
-            parameters["rows"] = config.max_results or 10000000
+            parameters["rows"] = config.max_results or 10_000_000
             # Check if rows param is 0 for backwards compatibility. Before
             # Solr 4 'rows = 0' meant that there is no limitation. Solr 4
             # always expects a rows param > 0 though:
             # http://wiki.apache.org/solr/CommonQueryParameters#rows
             if parameters["rows"] == 0:
-                parameters["rows"] = 10000000
+                parameters["rows"] = 10_000_000
             logger.debug(
                 'falling back to "max_results" (%d) without a "rows" '
                 "parameter: %r (%r)",
@@ -88,6 +88,20 @@ class Search(object):
                 query,
                 parameters,
             )
+        else:
+            try:
+                rows = int(parameters["rows"])
+            except ValueError:
+                # probably invalid, but we'll let solr complain about it
+                pass
+            else:
+                if rows > MAX_RESULTS_SUPPORTED_BY_SOLR:
+                    logger.warning(
+                        'The "rows" parameter was set to "%s" but automatically limited '
+                        ' to %s to avoid a Solr java.lang.NumberFormatException',
+                        parameters["rows"], MAX_RESULTS_SUPPORTED_BY_SOLR
+                    )
+                    parameters["rows"] = MAX_RESULTS_SUPPORTED_BY_SOLR
         if getattr(config, "highlight_fields", None):
             if parameters.get("hl", "false") == "true" and "hl.fl" not in parameters:
                 parameters["hl"] = "true"
@@ -111,12 +125,6 @@ class Search(object):
             field = schema.get(index, None)
             if field is None or not field.stored:
                 logger.warning('sorting on non-stored attribute "%s"', index)
-        if "rows" in parameters and parameters["rows"] > 1000000000:
-            logger.warning(
-                'The "rows" parameter was set to "%s" and automatically limited to 1000000000 to avoid a Solr java.lang.NumberFormatException',
-                parameters["rows"],
-            )
-            parameters["rows"] = 1000000000
         response = connection.search(q=query, **parameters)
         results = SolrResponse(response)
         response.close()