+++ CHANGES.rst
@@ -4,7 +4,8 @@ Changelog
 9.1.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Limit `rows` to 1000000000; higher values can cause a NumberFormatException in solr.
+  [tisto, davisagli]
 
 
 9.1.0 (2023-03-07)