+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 9.1.1 (unreleased)
 ------------------
 
+- Fix deserialization of image_scales metadata when there is no value
+  [mpeeters]
+
 - Limit `rows` to 1000000000; higher values can cause a NumberFormatException in solr.
   [tisto, davisagli]